1.0.2
Fixed a bug with Clear Highlighting on Workshop Exit that made it not work at all, nor would the settings change save.
Greatly improved the layer duplication code. Should be substantially faster, and will no longer back up the script queue.
WS+ will now post a warning pop-up if you forget to update Workshop Framework to the version it requires.
Added Layer Handles.
Layer Handles allow picking up and even cloning a layer.
Layer Handles can be built in workshop mode, or hotkeyed with MCM.
When a Layer Handle is built, it will switch to a version with the active layer number floating over it. It will then attach itself to all items on the current layer.
The Layer Handle model is special. It has a square around the number that will highlight red when you should NOT move it, and flash green once it’s OK for you to move it again.
There are currently two times when you shouldn’t move it:
After you first build it and it is attaching all items (you will see the items flash blue as they are attached). Moving during this will cause some items not to be attached.
When you are cloning the layer handle. Moving during this will cause some items to not clone in the positions they were at when you pressed the button as it takes a few seconds to copy all of the coordinates.
WARNING - It is not recommended to pick up the Default Layer handle unless you had scrapped the entire settlement. Any objects that aren’t buildable in Workshop Mode will disappear when you pick up the layer handle. This means scrappable objects or things you can’t built yourself, such as items placed by Sim Settlements City Plan. This is an engine limitation.
When you pick up a Layer Handle, all of the items that you will be moving will highlight green, similar to the highlight that happens when you group select in workshop mode.
WARNING - Do NOT store the layer handle. Doing so will cause all attached items to be scrapped or stored in the workbench. Once a Layer Handle exists in a settlement, it should never be removed. Layer Handles will autohide themselves when you exit workshop mode.
Notes when cloning a Layer Handle:
It will attempt to clone all items attached to the layer handle where they are at the time you pressed the Clone Key. Wait for the layer handle status to turn green before moving.
If the items are being cloned to the same layer the handle is for, those items will not be connected to the layer handle.
To connect new items to a layer handle, simply place the layer handle again, and it will connect everything on the layer to itself.
Undo/Redo will now support undoing/redoing cloning a group of objects via the Layer Handles or Duplicate Layer functionality.
Items created via bulk cloning (duplicate layer/cloning a layer handle) will become disconnected from the bulk undo possibility if you move or scrap them independently. This may change in the future, but tracking this is quite complex and can result in potential memory leaks if not done correctly.
Just like other Undo/Redo actions, the bulk undo/redo stack is cleared when you enter a new settlement.
WARNING: Undo/Redo does not currently work with moving objects via a Layer Handle, but that is planned for a future update.
Added new options to MCM/Holotape:
Clone Layer Handle Method: This let’s you configure which layer items attached to a layer handle should be cloned to. The options are as follows:
Active Layer [Default] - Items will be cloned to the active layer.
Source Layer - Items will be cloned to the same layer the originals are on.
New Layer - WS+ will attempt to create a new layer and clone to that new layer. If no new layer slots are available, the items will be cloned to the default layer.
Ask Me - You will be prompted to select how to handle the cloned objects.
Highlight Items When Layer Handle Held: Allows you to turn off the highlighting of items when picked up via a Layer Handle.
Added proper messages to Undo/Redo instead of relying on Debug messages. This is in preparation for a future update where support for Undo/Redo is added to Xbox.