Release 1.0.4
When you first load this patch, you will receive a series of Added/Remove messages. This is to fix your blueprint objects which will all be moved to the Blueprint Storage. If you haven’t renamed your blueprints, you will want to do so before installing this patch to avoid them stacking.
Layers added to the default layer will no longer be flashed purple. This means that players not using the layers feature won’t have to see their items flashing.
Added two options to MCM/Holotape to control whether the purple flashing occurs on items, the first option affects items added to the default layer, and the second affects items added to any layer.
Items built before installing Workshop Plus will no longer be added to the default layer automatically. This will reduce overall system load, as doing so can be very intensive, especially in complex settlements. For most players - there is little advantage to having those items on the default layer anyway.
There is now an MCM/Layer Control Module “Add Previous Items to a Layer” tool you can run on demand.
Fixed an issue that could prevent items from being moved to a different layer when a layer is cleared or deleted.
Layers will now correctly be removed from the layer select menu when they are deleted.
After deleting a layer, if you chose to move the items on it to a different layer, that different layer will be made active instead of just the next one in the list. (For example, if you deleted layer 1 and moved the items to layer 4, layer 4 would become active. Previously, layer 2 would have become active, which was not intuitive).
Added an option to MCM/Holotape prevent the hotkey warning from firing when you press a WS+ hotkey outside of workshop mode.