v1.0.6
Fixed a bug in the Layer HUD positioning, where it would invert the vertical position you chose after you reloaded the game.
Fixed a bug in the Layer HUD auto-appearing, where it wouldn’t always appear when you entered a settlement if you had it configured to display even outside of workshop mode.
Dramatically improved the code for clearing workshop highlighting so that it will be less likely to spam the script queue with actions.
Clear Highlighting Outside WS Mode option now defaults to OFF. This will reduce script executions for people who aren’t making use of highlighting in the first place.
Added a new Free Build feature. This makes workshop mode believe you have unlimited resources and bypasses some Perk requirements on objects (exception appears to be shops).
Note that while this mode is on, you will receive no resources for scrapping.
This was enabled by a new engine change Bethesda made with the latest patch - likely in advance of the VR Workshops system.
Added an option to turn on Free Build for all settlements to the holotape and MCM.
Added a tool to turn Free Build on or off for the nearest settlement to the holotape.
Added a hotkey to MCM to toggle Free Build in the nearest settlement.